requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-multipart==0.0.6
langchain==0.0.314
langchain-core==0.1.1
langchain-community==0.0.10
langchain-ollama==0.0.1
langchain-neo4j==0.1.1
langchain-experimental==0.0.36
neo4j==5.13.0
ollama==0.1.5
python-dotenv==1.0.0
pdf2image==1.16.3
pymupdf==1.23.6
pytesseract==0.3.10
loguru==0.7.2
